requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.12
idna==3.4
importlib-metadata==4.8.3
prettytable==2.5.0
requests==2.27.1
typing-extensions==4.1.1
urllib3==1.26.15
wcwidth==0.2.12
yara-python==4.2.3
zipp==3.6.0
